backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
polars==1.19.0
openpyxl==3.1.5
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
pydantic-settings==2.7.0
python-dotenv==1.0.1
aiofiles==24.1.0
psycopg2-binary==2.9.10
asyncpg==0.30.0
